Break out server instead of types
One thing must be broken out to prevent circular imports, this seemed
nicer.
