+++ httpgrpc/server/server_test.go
@@ -1,4 +1,4 @@
-package httpgrpc
+package server
 
 import (
 	"bytes"
@@ -11,7 +11,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/weaveworks/common/httpgrpc/types"
+	"github.com/weaveworks/common/httpgrpc"
 	"github.com/weaveworks/common/user"
 	"google.golang.org/grpc"
 )
@@ -34,7 +34,7 @@ func newTestServer(handler http.Handler) (*testServer, error) {
 		URL:        "direct://" + lis.Addr().String(),
 	}
 
-	types.RegisterHTTPServer(server.grpcServer, server.Server)
+	httpgrpc.RegisterHTTPServer(server.grpcServer, server.Server)
 	go server.grpcServer.Serve(lis)
 
 	return server, nil
